Rename routes dir to pages
