+++ src/pages/splash.vue
