+++ src/router/index.js
@@ -1,11 +1,11 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 
-import Splash from '../routes/splash.vue'
-import WardLeaderList from '../routes/ward-leader-list.vue'
-import WardLeader from '../routes/ward-leader.vue'
-import CityMap from '../routes/city-map.vue'
-import ContentPage from '../routes/content-page.vue'
+import Splash from '../pages/splash.vue'
+import WardLeaderList from '../pages/ward-leader-list.vue'
+import WardLeader from '../pages/ward-leader.vue'
+import CityMap from '../pages/city-map.vue'
+import ContentPage from '../pages/content-page.vue'
 
 Vue.use(VueRouter)
